requirements.txt
asgiref==3.7.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-nested-admin==4.0.2
pillow==10.2.0
psycopg2==2.9.9
python-monkey-business==1.0.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
